requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
anthropic>=0.40
voyageai>=0.3
google-api-python-client>=2.120
rapidfuzz>=3.6
pydantic>=2.6
python-dotenv>=1.0
numpy>=1.26
